+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ### Changed
 
 - `Raxx.set_header/2` now raises when setting host headers.
+- `ArgumentError` is raised instead of `RuntimeError` in cases of bad headers and body content.
 
 ## [0.16.1](https://github.com/CrowdHailer/raxx/tree/0.16.1) - 2018-09-19